refactor foundation workspace router
